+++ source/components/sidenav-simple/index.js
@@ -1,1 +1,2 @@
 import "./sidenav-simple.less";
+import "./sidenav-simple.js";